Ensure LFS is used on 32-bit compiles
I noticed file-verify.t failures on 32-bit compiles, and it's due to
not using the correct 64-bit LFS sizing. This means off_t is 32-bit,
rather than the 64-bit we expect.

Just set _LARGEFILE_SOURCE and _FILE_OFFSET_BITS=64 unconditionally,
and remove it from the two tests where we already used it.

+++ test/file-verify.c
@@ -46,7 +46,8 @@ static int verify_buf(void *buf, size_t size, off_t off)
 	ptr = buf;
 	for (i = 0; i < u_in_buf; i++) {
 		if (off != *ptr) {
-			fprintf(stderr, "Found %u, wanted %lu\n", *ptr, off);
+			fprintf(stderr, "Found %u, wanted %llu\n", *ptr,
+					(unsigned long long) off);
 			return 1;
 		}
 		ptr++;
